@@ -4,8 +4,8 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'image/user_image_widget.dart';
 
 class UserGridTile extends StatelessWidget {
-  final String userImageUrl;
-  final String userName;
+  final String? userImageUrl;
+  final String? userName;
   final Function()? onItemClicked;
 
   const UserGridTile(
@@ -23,20 +23,17 @@ class UserGridTile extends StatelessWidget {
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           UserImageWidget(
-              height: 180,
-              imageUrl: userImageUrl,
+              height: 150,
+              imageUrl: userImageUrl ?? "",
               borderRadius: const BorderRadius.only(
                   topRight: Radius.circular(12),
                   bottomRight: Radius.circular(12))),
           5.verticalSpaceFromWidth,
           Text(
-            userName,
+            userName ?? "",
             style: context.titleMedium,
             maxLines: 2,
             overflow: TextOverflow.ellipsis,
-          ),
-          const SizedBox(
-            height: 6,
           ),
         ],
       ),
