@@ -4,11 +4,13 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 class GridWidget extends StatelessWidget {
   final int itemsCount;
   final Widget? Function(BuildContext context, int index) itemBuilder;
+  final EdgeInsets? contentPadding;
 
   const GridWidget({
     super.key,
     required this.itemsCount,
     required this.itemBuilder,
+    this.contentPadding
   });
 
   @override
@@ -19,7 +21,7 @@ class GridWidget extends StatelessWidget {
         mainAxisSpacing: 8.w,
         crossAxisSpacing: 8.w,
       ),
-      padding: const EdgeInsets.all(8).w,
+      padding: contentPadding ?? EdgeInsets.zero,
       itemCount: itemsCount,
       itemBuilder: itemBuilder,
     );
