@@ -6,6 +6,7 @@
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i687;
 
+import 'package:domain/repositories/user_repository.dart' as _i988;
 import 'package:domain/usecases/user/get_users_usecase.dart' as _i977;
 import 'package:injectable/injectable.dart' as _i526;
 
@@ -13,6 +14,7 @@ class DomainPackageModule extends _i526.MicroPackageModule {
 // initializes the registration of main-scope dependencies inside of GetIt
   @override
   _i687.FutureOr<void> init(_i526.GetItHelper gh) {
-    gh.factory<_i977.GetUsersUseCase>(() => _i977.GetUsersUseCase());
+    gh.factory<_i977.GetUsersUseCase>(
+        () => _i977.GetUsersUseCase(gh<_i988.UserRepository>()));
   }
 }
