@@ -1,5 +1,5 @@
 import 'package:domain/models/user.dart';
 
 abstract class UserRepository{
-  List<User> getUsers();
+  Future<UserListResponse> getUsers();
 }
