@@ -2,14 +2,16 @@ import 'package:design/components/api_error_widget.dart';
 import 'package:design/components/list/users_grid_widget.dart';
 import 'package:design/components/user_grid_tile.dart';
 import 'package:design/extensions/resources_ext.dart';
-import 'package:features/home/bloc/home_cubit.dart';
 import 'package:features/uimodels/user/user_model.dart';
 import 'package:flutter/material.dart';
 import 'package:auto_route/auto_route.dart';
 import 'package:auto_route/annotations.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get_it/get_it.dart';
 import 'package:design/gen/assets.gen.dart';
+
+import 'cubit/home_cubit.dart';
 
 @RoutePage()
 class HomeScreen extends StatefulWidget {
@@ -46,7 +48,7 @@ class _HomeScreenState extends State<HomeScreen> {
         child: _usersListBuilder(),
       ),
       floatingActionButton: FloatingActionButton(
-        tooltip: 'Increment',
+        tooltip: 'Add new',
         onPressed: () {},
         child: const Icon(Icons.add),
       ), // This trailing comma makes auto-formatting nicer for build methods.
@@ -59,16 +61,14 @@ class _HomeScreenState extends State<HomeScreen> {
         if (state is HomeLoadSuccess) {
           return _usersList(state.users);
         } else if (state is HomeLoadError) {
-          return SliverFillRemaining(
-              hasScrollBody: false,
-              child: ApiErrorWidget(
-                errorText: state.message,
-                onRetry: () {
-                  BlocProvider.of<HomeCubit>(context).loadUsers();
-                },
-              ));
+          return ApiErrorWidget(
+            errorText: state.message,
+            onRetry: () {
+              BlocProvider.of<HomeCubit>(context).loadUsers();
+            },
+          );
         } else {
-          return  Center(
+          return Center(
             child: Assets.images.fire.asImage(),
           );
         }
@@ -77,13 +77,14 @@ class _HomeScreenState extends State<HomeScreen> {
     );
   }
 
-  _usersList(List<UserModel> users) {
-    GridWidget(
+  Widget _usersList(List<UserModel> users) {
+   return GridWidget(
+     contentPadding: const EdgeInsets.only(top: 12, bottom: 60, left: 12, right: 12).w,
       itemsCount: users.length,
       itemBuilder: (context, index) {
         return UserGridTile(
-          userImageUrl: users[index].photo,
-          userName: users[index].name,
+          userImageUrl: users[index].image,
+          userName: users[index].firstName,
         );
       },
     );
