@@ -1,19 +1,26 @@
+import 'package:domain/models/user.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+
 part 'user_model.freezed.dart';
 
 @freezed
 class UserModel with _$UserModel {
   const factory UserModel({
-    required int id,
-    required String name,
-    required String company,
-    required String username,
-    required String email,
-    required String address,
-    required String zip,
-    required String state,
-    required String country,
-    required String phone,
-    required String photo,
+    required int? id,
+    required String? firstName,
+    required String? lastName,
+    required String? email,
+    required String? phone,
+    required String? image,
   }) = _UserModel;
+
+  factory UserModel.from(User user) {
+    return UserModel(
+        id: user.id,
+        firstName: user.firstName,
+        lastName: user.lastName,
+        email: user.email,
+        phone: user.phone,
+        image: user.image);
+  }
 }
