@@ -8,11 +8,12 @@
 // coverage:ignore-file
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:data/di/data_injector.module.dart' as _i3;
-import 'package:domain/di/domain_injector.module.dart' as _i4;
-import 'package:features/di/features_injector.module.dart' as _i5;
+import 'package:data/di/data_injector.module.dart' as _i4;
+import 'package:domain/di/domain_injector.module.dart' as _i5;
+import 'package:features/di/features_injector.module.dart' as _i6;
 import 'package:get_it/get_it.dart' as _i1;
 import 'package:injectable/injectable.dart' as _i2;
+import 'package:network/di/network_injector.module.dart' as _i3;
 
 extension GetItInjectableX on _i1.GetIt {
 // initializes the registration of main-scope dependencies inside of GetIt
@@ -25,9 +26,10 @@ extension GetItInjectableX on _i1.GetIt {
       environment,
       environmentFilter,
     );
-    await _i3.DataPackageModule().init(gh);
-    await _i4.DomainPackageModule().init(gh);
-    await _i5.FeaturesPackageModule().init(gh);
+    await _i3.NetworkPackageModule().init(gh);
+    await _i4.DataPackageModule().init(gh);
+    await _i5.DomainPackageModule().init(gh);
+    await _i6.FeaturesPackageModule().init(gh);
     return this;
   }
 }
