@@ -1,5 +1,6 @@
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
+import 'package:network/di/network_injector.module.dart';
 import 'injector.config.dart';
 
 final getIt = GetIt.instance;
@@ -8,6 +9,6 @@ final getIt = GetIt.instance;
     includeMicroPackages: true,
     initializerName: 'init',
     preferRelativeImports: true,
+    externalPackageModulesBefore: [ExternalModule(NetworkPackageModule)],
     asExtension: true)
-
 Future<void> configureDependencies() async => getIt.init();
